Update admin page
Link to the user function
