+++ src/components/Admin/index.js
@@ -1,77 +1,31 @@
 
 import React, {Component} from 'react';
+import { Switch, Route, Link } from 'react-router-dom';
 import { compose } from 'recompose';
 
 import { withFirebase } from '../Firebase';
 import { withAuthorization, withEmailVerification } from '../Session';
+import UserList, {UserItem} from './UserList';
 
 import * as ROLES from '../../constants/roles';
+import * as ROUTES from '../../constants/routes';
 
-class AdminPage extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      loading: false,
-      users: [],
-    };
-  }
-  componentDidMount() {
-    this.setState({ loading: true });
-// example of get user
-//.on method from .users trigger when users is update to re render the Admin Page
-    this.props.firebase.users().on('value', userssnapshot => {
-      const usersObject = userssnapshot.val();
-      //Restructure object users as list
-      const usersList = Object.keys(usersObject).map(key => ({
-        ...usersObject[key],
-        uid: key,
-      }));
-
-      this.setState({
-        users: usersList,
-        loading: false,
-      });
-    });
-  }
-
-  componentWillUnmount() {
-    this.props.firebase.users().off();
-  }
-
-  render() {
-    const { users, loading } = this.state;
-
-    return (
+const AdminPage = () =>(
       <div>
         <h1>Admin</h1>
         <p>
           The Admin Page is accessible by every signed in Admin User.
         </p>
-        {loading && <div>Loading ...</div>}
-
-        <UserList users={users} />
+        {/*User to route ADMIN to a specific user detail*/}
+        <Switch>
+          <Route exact path={ROUTES.ADMIN_DETAILS} component={UserItem} />
+          <Route exact path={ROUTES.ADMIN} component={UserList} />
+        </Switch>
       </div>
-    );
-  }
-}
-const UserList = ({ users }) => (
-  <ul>
-    {users.map(user => (
-      <li key={user.uid}>
-        <span>
-          <strong>ID:</strong> {user.uid}
-        </span>
-        <span>
-          <strong>E-Mail:</strong> {user.email}
-        </span>
-        <span>
-          <strong>Username:</strong> {user.username}
-        </span>
-      </li>
-    ))}
-  </ul>
 );
+
+
+
 /*
 condtion check for role
  */
